+++ app/src/main/java/com/example/quickbuyapp/Database/CartDataSource.kt
@@ -13,7 +13,7 @@ interface CartDataSource {
     fun countItemInCart(uid: String): Single<Int>
 
 
-    fun sumPrice(uid: String): Single<Long>
+    fun sumPrice(uid: String): Single<Double>
 
 
     fun getItemInCart(productId:String,uid: String): Single<CartItem>
